Fixing discord bot + update to DJS12
