+++ commands_to_update/7daysfree.js
