+++ commands_to_update/kick.js
