+++ commands_to_update/spamcfg.js
