+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "brexite_but_as_a_bot",
-  "version": "3.1.1",
+  "version": "4.0.0",
   "description": "Added addreply",
   "main": "index.js",
   "scripts": {
@@ -11,7 +11,7 @@
   "dependencies": {
     "@discordjs/uws": "^10.149.0",
     "canvas": "^2.6.1",
-    "discord.js": "^11.6.4",
+    "discord.js": "^12.4.0",
     "express": "^4.17.1",
     "fs": "^0.0.2",
     "ms": "^2.1.2",
@@ -22,6 +22,6 @@
     "ytdl-core": "^2.1.1"
   },
   "engines": {
-    "node": "10.15.3"
+    "node": "16.9.1"
   }
 }
