--- txt/serverdata.json
@@ -1,37 +1,0 @@
-{
-	"serverData": {},
-	"280423419722465281": {
-		"bannedlords": [
-			"140175075314892800",
-			"164734460216999936"
-		],
-		"whitelist": [
-			"612022828782452746"
-		]
-	},
-	"644551231020204062": {
-		"bannedlords": [
-			"140175075314892800",
-			"164734460216999936"
-		],
-		"whitelist": [
-			"680379637695774723",
-			"644648782226522153",
-			"644638330587316264",
-			"644630452405272592",
-			"644631203508387859"
-		]
-	},
-	"696515024746709003": {
-		"bannedlords": [
-			"140175075314892800",
-			"164734460216999936"
-		],
-		"whitelist": [
-			"696740477533421618",
-			"696515024746709006",
-			"696538508227248178",
-			"697548272192979074"
-		]
-	}
-}
